requirements.txt
openai
pinecone-client
langchain
langchain-openai
langchain-pinecone
python-dotenv
pypdf
streamlit
tiktoken